refactor: split development verifier package
